fix: improve test for CI
